# BackEnd/docs/swagger.yaml
basePath: /
definitions:
...
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        format: int64
        type: integer
    type: object
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
  web_doscom_internal_database_model.PengurusPatch:
    properties:
      divisi:
        type: string
      email:
        type: string
      period:
        maxLength: 50
        type: string
      position:
        description: condition just for koor and super_admin
        type: string
      sosmed:
        type: string
      url_asset:
        $ref: '#/definitions/multipart.FileHeader'
    type: object
  web_doscom_internal_database_model.PengurusResponse:
    properties:
      divisi:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      period:
        type: string
      photo_url:
        type: string
      position:
        type: string
      sosmed:
        type: string
    type: object
...
info:
  contact: {}
  description: API Documentation Web Doscom
  title: Web Doscom API
  version: "1.0"
paths:
...
  /api/v1/pengurus/:
    get:
      consumes:
      - application/json
      description: Mengambil semua data penguru
      parameters:
      - description: Filter by divisi (optional)
        in: query
        name: divisi
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.PengurusResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all Pengurus
      tags:
      - Pengurus
    post:
      consumes:
      - multipart/form-data
      description: Create pengurus with profile picture upload to MinIO. All files
        uploaded to MinIO bucket.
      parameters:
      - description: Profile Picture (image file)
        in: formData
        name: file
        required: true
        type: file
      - description: User ID (optional, auto-filled for non-admin)
        in: formData
        name: id_user
        type: integer
      - description: Email (optional, auto-filled for non-admin)
        in: formData
        name: email
        type: string
      - description: 'Divisi - Valid values: bph, pemro, jaringan, medcrev, data'
        in: formData
        name: divisi
        required: true
        type: string
      - description: Name (2-150 characters)
        in: formData
        name: name
        required: true
        type: string
      - description: 'Position - Valid values: ketum, sdm, pr, pm, pm_ang, sekum,
          bendum, sek_ang, ben_ang, kor_pemro, kor_jaringan, kor_medcrev, kor_data,
          anggota, pemro_ang, jaringan_ang, medcrev_ang, data_ang'
        in: formData
        name: position
        required: true
        type: string
      - description: 'Social Media Platform (optional) - Valid values: instagram,
          linkedin, github. Leave empty for default ''instagram'''
        in: formData
        name: sosmed
        type: string
      - description: Period (YYYY-MM-DD format)
        in: formData
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.PengurusResponse'
        "400":
          description: Validation error - check divisi, position, or sosmed values
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new pengurus
      tags:
      - Pengurus
  /api/v1/pengurus/{id}:
    delete:
      consumes:
      - application/json
      description: Delete data pengurus berdasarkan ID
      parameters:
      - description: Pengurus ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.PengurusResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Pengurus
      tags:
      - Pengurus
    get:
      consumes:
      - application/json
      description: Mengambil data pengurus berdasarkan ID
      parameters:
      - description: Pengurus ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.PengurusResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Pengurus by ID
      tags:
      - Pengurus
    put:
      consumes:
      - application/json
      description: Update data pengurus berdasarkan ID
      parameters:
      - description: Pengurus ID
        in: path
        name: id
        required: true
        type: integer
      - description: Pengurus info
        in: body
        name: pengurus
        required: true
        schema:
          $ref: '#/definitions/web_doscom_internal_database_model.PengurusPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.PengurusResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Pengurus
      tags:
      - Pengurus
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"